Refactor: entity 설계 수정 - Order 요구사항 VO로 묶음 - entity에서 service의 dto 의존 제거(OrderItemCommand -> OrderItemInfo)
