+++ src/main/java/com/fhsh/daitda/order/domain/entity/Order.java
@@ -5,10 +5,11 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-import com.fhsh.daitda.order.application.command.OrderItemCommand;
+import com.fhsh.daitda.order.domain.vo.OrderItemInfo;
 import com.fhsh.daitda.order.domain.enums.OrderStatus;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
+import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
@@ -37,10 +38,10 @@ public class Order {
 	@OneToMany(mappedBy = "order", cascade = CascadeType.PERSIST, orphanRemoval = true)
 	private List<OrderItem> orderItems = new ArrayList<>();
 
-	private LocalDateTime deadlineAt;
 	@Enumerated(EnumType.STRING)
 	private OrderStatus orderStatus;
-	private String requestMessage;
+	@Embedded
+	private OrderRequest orderRequest;
 
 	public void addOrderItem(OrderItem item) {
 		orderItems.add(item);
@@ -48,12 +49,12 @@ public class Order {
 	}
 
 	// Order 와 OrderItem 저장 확인 위해 간단히, 추후 수정 예정
-	public static Order create(List<OrderItemCommand> itemInfos) {
+	public static Order create(List<OrderItemInfo> itemInfos) {
 		Order order = new Order();
 
 		order.orderStatus = OrderStatus.CREATED;
 
-		for (OrderItemCommand info : itemInfos) {
+		for (OrderItemInfo info : itemInfos) {
 			OrderItem item = OrderItem.create(
 				new OrderProduct(info.productId(), info.productName(), info.quantity())
 			);
