+++ src/main/java/com/fhsh/daitda/order/domain/entity/OrderRequest.java
@@ -1,0 +1,12 @@
+package com.fhsh.daitda.order.domain.entity;
+
+import java.time.LocalDateTime;
+
+import jakarta.persistence.Embeddable;
+
+@Embeddable
+record OrderRequest(
+	LocalDateTime deadlineAt,
+	String requestMessage
+) {
+}
